removes TryResponse interface
