+++ commons-api/src/main/java/cern/molr/commons/api/type/trye/Failure.java
@@ -5,6 +5,8 @@
 package cern.molr.commons.api.type.trye;
 
 import cern.molr.commons.api.type.either.Left;
+import com.fasterxml.jackson.annotation.JsonIgnore;
+import com.fasterxml.jackson.annotation.JsonProperty;
 
 /**
  * Refer to {@link Try}
@@ -18,6 +20,18 @@ public class Failure<T> extends Left<Throwable, T> implements Try<T> {
 
     public Failure(Throwable throwable) {
         super(throwable);
+    }
+
+    @Override
+    @JsonProperty("throwable")
+    public Throwable getThrowable() {
+        return leftValue;
+    }
+
+    @Override
+    @JsonIgnore
+    public T getResult() {
+        return null;
     }
 
 }
