+++ commons-api/src/main/java/cern/molr/commons/api/type/trye/Success.java
@@ -5,6 +5,8 @@
 package cern.molr.commons.api.type.trye;
 
 import cern.molr.commons.api.type.either.Right;
+import com.fasterxml.jackson.annotation.JsonIgnore;
+import com.fasterxml.jackson.annotation.JsonProperty;
 
 /**
  * Refer to {@link Try}
@@ -18,6 +20,18 @@ public class Success<T> extends Right<Throwable, T> implements Try<T> {
 
     public Success(T successValue) {
         super(successValue);
+    }
+
+    @Override
+    @JsonIgnore
+    public Throwable getThrowable() {
+        return null;
+    }
+
+    @Override
+    @JsonProperty("success")
+    public T getResult() {
+        return this.rightValue;
     }
 
 }
